Move all moderation commands from main bot script to cog
